permission_broker: Look up udev device from devnode using st_rdev
To lookup a /sys device path from a devnode path, we previously
enumerated all devices and looked for a match. This took O(N) where N is
the number of devices in sysfs, and was fairly costly as sysfs access
are more expensive than a typical filesystem.

To make it worse, various parts of Chrome would try to check for access
on the device as it enumerates them. This means O(N^2) sysfs operations;
this would hog for as much as 3 seconds of CPU time when the login
screen was shown.

There is a trick used by sd_device_new_from_devname that makes use of
stat() and st_rdev to do the lookup with O(1) syscall. As we are using
the old udev API instead of sd_device API, we will need to implement
this by ourselves.

The new method lowers the overall cost from O(N^2) total to O(N).

BUG=b:336951771
TEST=USB FIDO2 security key can be used on websites.
TEST=USB FIDO2 security key can be accessed from ARCVM.
TEST=permission_broker consumes <100ms CPU time when login screen is shown.

Change-Id: I9187fd426a62ef705c1ec2bae74273f887c42695
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5742977
Commit-Queue: Tatsuyuki Ishi <ishitatsuyuki@google.com>
Reviewed-by: maciek swiech <drmasquatch@google.com>
Tested-by: Tatsuyuki Ishi <ishitatsuyuki@google.com>