cryptohome: Minimal patch to allow pinweaver to be used with passwords
This keeps the feature completely disabled by default, via what is
currently a compile time flag. If we decided it would be a good idea to
roll it out we'd almost certainly need to make it a more dynamic flag.
We'd also need to add support for migrating existing passwords, which is
actually the messier aspect of the operation.

BUG=b:172221456
TEST=FEATURES=test emerge-$BOARD cryptohome
TEST=tast run cryptohome.* (with feature enabled manually)

Change-Id: Ic704224e047135c97a981d27d601c873a76b9147
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5754664
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>