resourced: rework Dynamic EPP initialization
Allow enabling Dynamic EPP straight from arch::init().
Additionally fix a race condition in spawning the auto_epp_main
by introducing a conditional-variable pattern to park the
coroutine if the feature is disabled.

Do not use feature::is_feature_enabled() and instead rely on a global
atomic variable to determine the feature state to save time on a string
hash generation which is needed in the hashmap search at the memory cost
of std::sync::atomic::AtomicBool. Though it should be noted that the
auto_epp_main function itself has a lot of space for optimization.

BUG=b:265774037
TEST=cros workon --board=rex start chromeos-base/resourced
TEST=cros build-packages --board=rex chromeos-base/resourced
TEST=cros deploy $DUT_IP chromeos-base/resourced
TEST=ssh dut.office "restart resourced"
TEST=ssh dut.office "grep resourced.*Dynamic /var/log/messages"

Change-Id: Icd33de141cb9038259399c0338abc04b855ca470
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5756724
Reviewed-by: David Stevens <stevensd@chromium.org>
Tested-by: Stanisław Kardach <skardach@google.com>
Commit-Queue: Stanisław Kardach <skardach@google.com>