vm_tools: concierge: exit if the KVM device is missing on Flex
http://go/crash-_ba5c2c06cf6cfd69 is consistently one of Flex's top
crashes. We do not look for some real bugs hidden amongst the much more
common case of KVM not being available. The only thing that will fail if
KVM is missing is crostini, which is not a big concern for Flex
users. So, we change vm_concierge's startup script to stop gracefully if
the device is Flex and the KVM device is missing.

BUG=b:228857973
TEST=emerge-reven vm_host_tools
scp vm_concierge.conf dut
<Start vm_concierge and observe that  it exits gracefully on Flex device without KVM>
TEST=emerge-coral vm_host_tools
scp vm_concierge.conf dut
<Start vm_concierge and observe that it runs on Non-Flex device>

Change-Id: Ib3b2b4d312327e3348e16a21d9830f3070635ffb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5758056
Tested-by: Andre Braga <andrebraga@google.com>
Commit-Queue: Andre Braga <andrebraga@google.com>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>