featured: Add a new highres off experiment enabled on all boards
Previously we enabled highres timers only on Octopus since we were
seeing instabilities on a few other boards (b/356629277). It is not
clear why and this experiment is restart the effort since some time has
passed. Further, we could disable the feature on offending boards (as
mentioned in b/356629277#comment7) but first let us see again if the
feature is still causing issues and that it is not cross-talk or some
other anomally that happened earlier during the year.

BUG=b:356629277
TEST=none

Change-Id: Ide4afae8aed87dedda2869fde18a812cca78eb0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5761325
Tested-by: Joel Fernandes <joelaf@google.com>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Joel Fernandes <joelaf@google.com>
Commit-Queue: ChromeOS Auto Runner <chromeos-auto-runner@chromeos-bot.iam.gserviceaccount.com>