shill: wifi: Prevent disconnection when disconnected with the previous pending service
There is a delay between supplicant emits CurrentBSSChanged signal and
shill receives and handles the signal. shill disconnects with the
current and pending services when connecting to a new WiFi service.
Suppose shill is connecting to service1, if shill requests connecting to
service2 before handling the CurrentBSSChanged signal BSS1 -> null,
shill disconnects with service2 because it thinks the affected service
is the pending service (service2). This may generate connection failure
if the user switches networks frequently.

Find the affected service based on the supplicant BSS when handling a
CurrentBSSChanged signal, so that the pending service is not
disconnected.

BUG=b:331136457, b:339266878
TEST=FEATURES=test emerge-$BOARD shill
TEST=Open a WiFi quick setting menu, switch between two networks fast,
there is no connection failure.

Change-Id: I7bebdaf23560d3d5ccd5d95c55cea72f8135be9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5766957
Tested-by: Kaidong Wang <kaidong@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Kaidong Wang <kaidong@chromium.org>