cryptohome: Make pinweaver passwords a feature
The current code guards the feature with a compile time constant which
is simple but does require a re-compile in order to make use of and
test. It also would be more difficult to roll out since it would make
rollback more difficult, also requiring a fresh build.

This instead converts it to a feature flag which is much easier to set
and unset.

BUG=b:172221456
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I93b59a8002fb30cf7728e1814a475c706f627f50
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5768962
Reviewed-by: Zi Lin <lziest@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>