shill: Set up routing policies at Network::Start()
Previous to this patch, shill set up the routing policies for a Network
only after the Network is Connected. This patch changes this behavior
and set up the routing policies once when Network::Start() is called.

Because the Network does not have any address and is not assigned a
priority at this stage, the routing policies won't have the address
related rules and will be installed at default priority (32). An example
would be:

1320:   from all fwmark 0x3ee0000/0xffff0000 lookup 1006
1320:   from all oif eth-v0 lookup 1006
1320:   from all iif eth-v0 lookup 1006

These rules (particularly, the oif rule) is needed for sending packets
needed for network provisioning, mostly importantly Neighbor
Advertisement, as fe80::/64 route is added into per-interface table
because of accept_ra_rt_table. Without these rules, the Neighbor
Solicitation from DHCPv6 server won't get our response and we won't be
able to receive the DHCPv6 Advertise.

BUG=b:350884946
TEST=unit
TEST=manual: deploy to atlas and verify rule installed on a
             non-Connected network.

Change-Id: I9fddd14ecc24335ba29fef37aa93e7852e31fcd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5772154
Commit-Queue: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Taoyu Li <taoyl@chromium.org>