shill: Add helper for array{dict} property in Manager
Add HelpRegisterDerivedKeyValueStores in Manager, which acts as a helper
for registering array{dict} D-bus properties in Manager. This function
is nearly identical to HelpRegisterDerivedKeyValueStore, except it uses
KeyValueStores, which is an alias for std::vector<KeyValueStore>, and it
does not support defining a setter function, similar to
HelpRegisterConstDerivedStrings.

Small formatting changes because https://crrev.com/c/5750481 didn't run
`cros format`.

BUG=None
TEST=FEATURES=test emerge-brya shill

Change-Id: I845533da695967e9e78147546ad62274e8571ced
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5785552
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>