ippusb_bridge: Use explicit Context
Instead of relying on the default GlobalContext, create an explicit
Context before opening a device and reuse it in other places that refer
to the same device.  This should not result in any behavior change, but
will allow subsequent CLs to use new APIs that aren't available on
GlobalContext.

BUG=b:359261398
TEST=tast run scanner.*; manual scanning

Change-Id: Ide17bc41923b24b14422b1323576120476dc3e03
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5793616
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>