resourced: Add MGLRU split generation experiment
Add experiment for testing splitting anon and file generations apart in
the MGLRU. The MGLRU change should make the swappiness tunable function
properly, so add it as a parameter for the experiment.

In addition to the experiment, game mode changes the swappiness value.
To make the two settings work together nicely, add a central place for
synchronizing swappiness settings and issuing dbus requests.

BUG=b:342531825
TEST=tast run DUT multivm.Lifecycle.host_with_bg_arc_vmmms

Change-Id: Ie75b362f023358ddd210f28174a16b6e0de1e1cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5782642
Tested-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Shin Kawamura <kawasin@google.com>
Commit-Queue: David Stevens <stevensd@chromium.org>