resourced: schedqos: support downgrading RT for Urgent state
Re-implement the SetRtForDisplayThreads Chrome experiment in
libqos. With that experiment, Chrome threads with type kCompositing
and kDisplayCritical in a foreground process would be scheduled with
realtime priority [1]. Those thread types correspond to the Urgent
thread state [2], so when the experimental flag is enabled, schedule
Urgent threads in a foreground process at realtime priority.

[1] https://source.chromium.org/chromium/chromium/src/+/main:base/threading/platform_thread_cros.cc;l=271-288;drc=07051ab04c8c75320330dfe237c315425abb4c0e
[2] https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/ash/schedqos/dbus_schedqos_state_handler.cc;l=352-355;drc=47baeac442ed1f5c1a1c961fe0c8da13bdb37bc8

Downgrading RT feature was removed from schedqos by
https://crrev.com/c/5693978 due to audio regression. This change revives
that support, makes it configurable on a ThreadState by ThreadState
basis, and applies it to Urgent.

Since downgrading RT is an experiment for now, we don't support loading
rt priority from chromeos-config for RtPriority::Foreground in this CL.

The initialization steps in service_main() is reordered because
SchedQoSContext need to be initialized after feature manager init.

The feature can be updated after resourced initialization. Updating
schedqos config will be supported by following CL.

BUG=b:358232020
TEST=cargo test
TEST=cargo test -p schedqos
TEST=tast run $DUT sched.ChromeQoS with \
--enable-features=SchedQoSOnResourcedForChrome,SetRtForDisplayThreads

Change-Id: Ia0d18784dbb31c7b5d63d957c6c47c4a5c3d2c24
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5783882
Reviewed-by: David Stevens <stevensd@chromium.org>
Commit-Queue: Shin Kawamura <kawasin@google.com>
Tested-by: Shin Kawamura <kawasin@google.com>