cryptohome: Fix crash on failed GetCertificate calls
The dbus proxy API does not seem to allow for the ErrorPtr* parameter to
be null, so code should not call it with nullptr. This wasn't caught by
tests because this function is just mocked out and so there was no
actual behavior that would try to write to the pointer.

BUG=b:359626424
TEST=added unit test case for this

Change-Id: I4c934fcd8cae798f0e1e0fcb4917b38f4a1e498a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5786916
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Auto-Submit: John Admanski <jadmanski@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>