libchromeos-rs: Add mount module
This is the combination of the two existing mount modules in flexor and
os_install_service, which will be removed in future commits. The one
noteworthy difference is that the constructor has been renamed from
`mount` to `new`, to satisfy clippy's aversion to
"self-named-constructors".

Combining these in preparation for adding a third user of mount:
chromeos-install.

BUG=b:356344778
TEST=CQ

Change-Id: Ic068b1bd2065259a7e7d362101100870a12e5e39
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5788842
Commit-Queue: Ted Brandston <tbrandston@google.com>
Tested-by: Ted Brandston <tbrandston@google.com>
Reviewed-by: Allen Webb <allenwebb@google.com>