shill: cellular: Handle optional IP types
Restructure GetApnTypes to return nullopt when the IpType is not
present. This will be useful when adding the optional ip type for
roaming.

BUG=b:359064950
TEST=Change of the the ip_types in modb to unknown and run
cellular_mobile_operator_info_tester to verify the LOG entry is printed,
and the IpType is set to the default IPv4.

Change-Id: Ie2846641909a6693cc88f78b6011e7db6dc89e46
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5789535
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Madhav <madhavadas@google.com>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>