ippusb_bridge: Refactor hotplug handling
The original hotplug handler functions on Context are deprecated and
replaced with a new HotplugBuilder.  Switch to this to get rid of the
deprecation warnings.  It may also address a crash that occasionally
happens when the device is unplugged.

In the process, also refactor UnplugHandler out into a separate hotplug
file so that we can later stop building this on platforms where it isn't
available.

BUG=b:302132427, b:359261398
TEST=tast run scanner.*; manual plug and unplug devices

Change-Id: Iead2d0e9046a9c5987e316e5547ade8a8680fbef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5793618
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>