secagentd: fileevents: add event coalescing
Add event coalescing to file events. The read or modify file events that
are accumulated over a batching interval will be coalesced into fewer
events for batching.

This is to reduce the number of messages that need to be egressed by the
system and to reduce the number of SHA256s performed.

BUG=b:343986520
TEST=unit testing

Change-Id: I5e3b26b1a6c6036fb528a15463749ed76092f2fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5789170
Reviewed-by: Aashay Shringarpure <aashay@google.com>
Reviewed-by: Anjali vinayak (Anji) <avinayak2530@gmail.com>
Reviewed-by: Princy Agrawal <princya@google.com>
Auto-Submit: Jason Ling <jasonling@chromium.org>
Commit-Queue: Jason Ling <jasonling@chromium.org>
Tested-by: Jason Ling <jasonling@chromium.org>