heartd: Preserve Intel PMT log until next heartd start up
Without this change, if the engineer delete the config file on the DUT,
the log file may be deleted immediately when the CleanUp function is
called. This will cause trouble and make it inconvenient to use.

Therefore, we check if we should remove the log file in next start up.

BUG=b:289023919
TEST=FEATURES=test emerge-rex chromeos-base/heartd

Change-Id: I69fe2c74b96eecab1e39ac159ec82d3cbd65e1e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5785429
Tested-by: Kerker Yang <kerker@chromium.org>
Commit-Queue: Kerker Yang <kerker@chromium.org>
Reviewed-by: Denny Huang <dennyh@google.com>