resourced: schedqos: add traverse() to ProcessMap
In later CL, we will add update_config() which updates the scheduler
settings of all processes and threads managed by schedqos.
ProcessMap::traverse() will be used for it.

This contains the refactoring that uses a mutable reference of an entry
in ProcessMap as ProcessContext instead of OccupiedEntry. This is
because HashMap does not provide API to get OccupiedEntry of all
entries. It is possible to get OccupiedEntry of all entries by calling
HashMap::entry() for all stored key, but it has more cost than
HashMap::iter_mut().

The refactoring ended up searching an entry twice in
ProcessMap::insert_or_update() (e.g. SimpleProcessMap calls
`self.entry()` and `self.get_mut()`. However it is inevitable due to
Rust lifetime constraint.

BUG=b:358232020
TEST=sudo cargo test -p schedqos

Change-Id: I500cf60c5b0a18cd49789fd1e4706955f0734f47
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5790170
Tested-by: Shin Kawamura <kawasin@google.com>
Commit-Queue: Shin Kawamura <kawasin@google.com>
Reviewed-by: David Stevens <stevensd@chromium.org>