fpstudy: Avoid unused linter warnings by using fake_use func
Introduce a fake_use function, so that the linter doesn't complain
about variables that appear unused. These variables are used inside
the benchmark function. Their reference can't be interpreted, since
the statement is encoded as a string. Even if it was not a string,
the real reference path is through the passed in globals variable.

BUG=b:343809394
TEST=python -m unittest discover -v -p '*_test.py'

Change-Id: I9f5ff894ae180efdeb888f0b8bff7b937cbc20da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5790558
Commit-Queue: Craig Hesling <hesling@chromium.org>
Reviewed-by: Firas Sammoura <fsammoura@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Tested-by: Craig Hesling <hesling@chromium.org>