camera: Add OpenDeviceError metric
This CL adds logic to send metrics when the camera is failed to be
opened and the value helps us understand whether the reason is that
camera is currently occupied. It also contains a logic to only send
such event once in a camera session.

BUG=b:360312017
TEST=Build successfully

Change-Id: I00bf341635da1e95e85373a7db19728964f84c32
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5795357
Reviewed-by: Shik Chen <shik@chromium.org>
Tested-by: Wei Lee <wtlee@chromium.org>
Commit-Queue: Wei Lee <wtlee@chromium.org>