shill: cellular: Enable tethering on KDDI, AT&T, Hanshin, EE
Enable tethering capability on more carriers. In this wave, we are
enabling it on AU by KDDI, AT&T US, Hanshin Cable(JP) and EE UK.

BUG=b:360391413
TEST=Tested on crota/AT&T and pujjoteen5/KDDI in the lab.
Verified that the UI shows the Hotspot menu, and was able to start a
hotspot session.
Note: For pujjoteen5, I had to temporarily remove the variant from the
blocklist in VariantSupportsTethering.

Change-Id: I22bba06db844e9865fc3270d28cdc3bb58c10c43
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5792414
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>