cryptohome: Replace a lot of NOTREACHED_IN_MIGRATION
The NOTREACHED log is being changed to a fatal error, and so I'm trying
to move the code away from using the transitional message.

I don't think any of these instances should actually be fatal, or
equivalent to a CHECK, and so I'm generally replacing them with
LOG(ERROR) messages and then continuing. At least a bunch of them are
clearly written with the expectation that execution will continue, with
code after the NOTREACHED, so in a lot of the cases I think this is in
line with the original intent.

There are a few instances I haven't touched because I think the code is
doing the opposite, expecting it to be a fatal error. It's not obviously
to me that this is the correct behavior, though, so for now I'm leaving
those untouched to be handled in a separate change.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I3e08d5390b9ff4de902199c25e43c9c3ecc498b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5794094
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>