cros-disks: Enhance RenameManager::StartRenaming
Enhanced RenameManager so that it can rename external drives that are
mounted by an in-kernel driver or by a FUSE mounter. This is done by
always running the renaming program with the `disk` group, and adding
the `fuse-exfat` or the `ntfs-3g` group as a supplementary group if
necessary.

BUG=b:358446133
TEST=cros_workon_make --board=$BOARD --test chromeos-base/cros-disks
TEST=tast run $DUT platform.CrosDisks*

Change-Id: Ia52867b3d7cfa404d2db8841bae88f1623fa867d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5795585
Commit-Queue: François Degros <fdegros@chromium.org>
Reviewed-by: Ben Reich <benreich@chromium.org>
Tested-by: François Degros <fdegros@chromium.org>