fpstudy: Fix FAR full hierarchy bootstrapping edge case
We are finally addressing and fixing the concern about including
invalid verification fingers in the bootstrapping sample space.
For example, if we are processing the combination of enrollment
user 1 against verification user 1, we would expect to only match
the enrollment finger with the remaining fingers of the user. In
the unfixed code, we would still consider the case where the
fingers are the same. For one reason or another, this actually
didn't seem to impact the FAR results from real or simulation
data, but it was a lingering concern.

Having this additional logic doesn't negatively impact performance
in any noticeable way. It actually seems to reduce bootstrap sampling
runtime by about 5%. I saw the total bootstrap sampling runtime drop
from about 7.4s to about 7s when taking 1000 bootstrap samples
sequentially (num_proc=None) and default SimulateEvalResults()
parameters.

There are two justifications for this improved performance:
1. Compared to the original commented out method for filtering the
   enrollment finger from the verification finger list, now all compute
   overhead was moved to the _init function.
2. It is possible that the minuscule overhead that it adds to select the
   filtered finger list is dwarfed by the possible large performance
   gains caused by reducing the number fingers in the list by one, thus
   reducing the subsequent loop iterations.
3. It is harder to get an accurate benchmark of this, since the
   optimizations seen in this _sample function are mainly escape style,
   where large chunks of computation may be bypassed based on random
   numbers. Furthermore, this change will only take effect when we
   happen to select the same user for enrollment and verification.

BUG=b:343809394
TEST=python -m unittest discover -v -p '*_test.py'
TEST=Traced the runtime with prints.

Change-Id: I87ddd16b9d69ccf36b0430d82f57279cfb7ab374
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5795360
Tested-by: Craig Hesling <hesling@chromium.org>
Reviewed-by: Firas Sammoura <fsammoura@google.com>
Commit-Queue: Craig Hesling <hesling@chromium.org>