shill: Network: Implement DHCPv6-PD callback
If the size of delegated prefix is at least /64, Network will take the
second address in the prefix as the IPv6 address of the host, and write
it into NetworkConfig, which will take preference over the address from
SLAAC (note that the expectation is that DHCPPD should be used together
with PIO P-flag, so no SLAAC address should coexist with PD).

The prefixes longer than /64 are ignored.

BUG=b:350884946
TEST=unit
TEST=manually verify prefix::2 configured on the interface
TEST=tast run $(DUT) network.DHCPPD*

Change-Id: Ibabc9702a7748de455d598f9b6777fa4cabb00e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5799446
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Zikai Chen <chenzikai@google.com>