arc: Get vbmeta digest value and set verified boot params
The value for vbmeta digest will be stored in a txt file in the
same directory as the system image. Retrieve the value from that
file and set all verified boot params in both the keymint and
remote provisioning contexts. In the case that the file is not
able to be read, the vb meta digest boot parameter will not be
set in CreateDeviceInfo.

board_specific_setup CL: https://crrev.com/i/7619063

BUG=b:350826299
BUG=b:350826304
TEST=USE="-cros-debug" cros_workon_make --test --board=dedede
chromeos-base/arc-keymint
TEST=add logs and verify expected values are used in CreateDeviceInfo
TEST=cros deploy and verify Play Store launches successfully

Change-Id: Ic476d3c67b9ae30eaf870fa4a65c960a02361593
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5809618
Tested-by: Luzanne Batoon <batoon@google.com>
Reviewed-by: Vaibhav Raheja <vraheja@chromium.org>
Auto-Submit: Luzanne Batoon <batoon@google.com>
Commit-Queue: Luzanne Batoon <batoon@google.com>