cryptohome: Only depend on 'fake_platform_features' from the test runner
This CL is a part of `platform2/featured` cleanup process.

The library `fake_platform_features` is linked from the library that is installed into the device, even though the content is only used from the unit test code.
This CL adjusts the build file and code to only depend on the library when building a test runner.

BUG=b:242920490
TEST=`emerge-hatch chromeos-base/cryptohome` succeeds.
TEST=`FEATURES=test emerge-hatch chromeos-base/cryptohome` succeeds.
TEST=After emerging the `fake_platform_features` to `/build/lib64`, `emerge-hatch chromeos-base/cryptohome` succeeds.
TEST=After emerging the `fake_platform_features` to `/build/lib64`, `FEATURES=test emerge-hatch chromeos-base/cryptohome` succeeds.

Change-Id: Ifbb0d9be09d056ef6c9c9ec13fde3b16468a73dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5802564
Tested-by: Soichiro Fujii <soichirof@google.com>
Reviewed-by: Takaya Saeki <takayas@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: Soichiro Fujii <soichirof@google.com>