odml: Adjust the empty mojom struct declaration
The empty CacheEmbeddingsResponse struct was declared with Native
attribute because the compile error message suggests that structs with
form like:
struct Foo;
needs to have the Native attribute. But this causes some linking
problems that are discovered afterwards. Turns out that the syntax
struct Foo {};
will work without native attribute, so changing to that.

BUG=b:358531135
TEST=build ok

Change-Id: I6c1237137ee802bd5d2ce8d2cb8ca6815706159a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5823020
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>