odml: Temporary remove some Stable from coral mojom
We want to modify some fields of the stable mojoms, but we don't want to
keep the backward compat parts as the mojom types aren't used in
anywhere in the codebase yet. Temporary remove the Stable attributes,
then in the next CL that modifies the types to what we want, we'll add
back the Stable attributes.

BUG=b:362443175
TEST=odml unittests

Change-Id: I4159fae1081809905e12e662056eca53283c3443
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5826173
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>