patchpanel: IPv6SourceEnforcement taking all IPv6 addresses
This patch modifies UpdateSourceEnforcementIPv6Prefix so that it allows
traffic from prefixes in all IPv6 addresses on the interface instead of
only the first one. This helps in networks where multiple PIOs are
announced.

Also modifies the interface of the function that it now takes the
original ipv6_addresses rather than the /64 prefixes.

BUG=b:361326673
TEST=unit

Change-Id: I71f0a56ad039ee849b9c964a33e3ad52436d984f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5824121
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Taoyu Li <taoyl@chromium.org>