shill: DHCPCDProxy: Send SIGALRM to dhcpcd 10 first in the cleanup
Currently when we disconnect a service, we both send a SIGALRM to dhcpcd
to release the lease, and send a SIGTERM to it to kill the process.
Since we are sending the SIGALRM from another dhcpcd process and sending
the SIGTERM from shill directly, these two operations form a race
condition. Normally SIGTERM comes earlier, so dhcpcd fails to send a
DHCPRELEASE message in the end and this causes tast network.DHCPHostname
to fail.

In this patch, before sending SIGTERM to kill the process, we also
directly send a SIGALRM from shill. This signal always comes first and
it helps releasing the lease and stopping the dhcpcd process. We still
use process_manager_->StopProcessAndBlock() because we want it to send
SIGKILL if the SIGALRM has failed to kill the process. Although
StopProcessAndBlock() still sends SIGTERM at the beginning, this is
ignored by dhcpcd, for dhcpcd will set its dhcpcd_exiting variable to
true after the first SIGALRM comes, and will ignore all subsequent
SIGTERM signals.

Note that since we still use dhcpcd 7 at the moment, this patch takes
no effect.

BUG=b:351710070
TEST=FEATURES=test emerge-cherry net-base
TEST=FEATURES=test emerge-cherry shill
TEST=tast run ${DUT} network.DHCPHostname

Change-Id: Ic0bc5a5f94b6d6955d9c423e01c459aba0855d7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5825600
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Zikai Chen <chenzikai@google.com>
Commit-Queue: Zikai Chen <chenzikai@google.com>