cryptohome: Ensure that the right pin policy is set
During migration of a pin VaultKeyset to USS and when the timeout pin
flag is enabled, a new pin is created. This pin is created based on the
timeout flag. This CL just checks that the policy bit in metadata is set
correctly.

There was another way to do this, where we set legacy pin schedule on
migration, but our eventual goal is to get everybody on the modern pin
schedule so that is why it was not adopted.

BUG=b:362712216
TEST=CQ Passes

Cq-Depend: chromium:5823070
Change-Id: I1d443efb7081c18e413e3140a210d53012f8d367
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5830350
Reviewed-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>