odml: Use GroupRequestPtr instead of ref as params
As GroupRequest isn't copyable, using ref as param makes it hard to
further bind to other callbacks. It is easier to just pass Ptr as param,
and move it to the response callback as well. Also update const
Response& fields to take the owned Response object, as it's easier for
binding.

BUG=b:358531135
TEST=odml unittests

Change-Id: Ic519eba45963926abe0d81cdac6cf97373940c51
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5833155
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>