mount-encrypted: do not use tpm for system key when runtime TPM selection enabled
When testing reven on chromebook, we would try to use TPM for system
key, since a TPM is present.
However, reven is compiled with USE flag |tpm_dynamic|, which is not
supported by the tlcl wrapper, we should not use the TPM for the system
key.
Fix the logic to make sure we are not using the TPM when running a reven
image on chromebooks.

BUG=b:359560664
TEST=Build a reven image with this change, check we can boot Nissa over USB.

Change-Id: I61802e27e620128b44469f6a31ae42b436341f9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5833288
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Nehemiah Dureus <ndureus@chromium.org>
Auto-Submit: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Nehemiah Dureus <ndureus@chromium.org>