shill: LegacyDHCPCDProxy: Send SIGALRM to dhcpcd 7 first in the cleanup
Currently when we disconnect a service, we both send via D-Bus to dhcpcd
to release the lease, and send a SIGTERM to it to kill the process.
These two operations form a race condition and may cause dhcpcd to fail
to send the RELEASE message.

In this patch, before sending SIGTERM to kill the process, we also
directly send a SIGALRM from shill. This signal always comes first and
it helps releasing the lease and stopping the dhcpcd process. We still
use process_manager_->StopProcessAndBlock() because we want it to send
SIGKILL if the SIGALRM has failed to kill the process. Although
StopProcessAndBlock() still sends SIGTERM at the beginning, this is
ignored by dhcpcd, for dhcpcd will set its dhcpcd_exiting variable (see
patch crrev.com/c/5831646) to true after the first SIGALRM comes, and
will ignore all subsequent SIGTERM signals.

BUG=b:351710070
TEST=FEATURES=test emerge-cherry shill
TEST=tast run ${DUT} network.DHCP*

Change-Id: I15ebfa001404b08a4f72857703d06ef362cbc8cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5831158
Tested-by: Zikai Chen <chenzikai@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Zikai Chen <chenzikai@google.com>