vm_tools: add verified boot parameters
During ARC Verified Boot, three parameters are used to establish
the root of trust and are included in the attestation certificate.
ChromeOS system properties are used to determine the values for
vbmeta device state and verified boot state. The vbmeta digest is
calculated when the ARCVM image is built and signed. The value is
written in a file which is retrieved in this code.

BUG=b:345550876
BUG=b:345556600
TEST=USE="-cros-debug" cros_run_unit_tests --board dedede --packages
chromeos-base/vm_host_tools
TEST=manually check that all 3 verified boot ro.boot params are set

Change-Id: I71654d64f27161795f367fe0b5a5c26fd54034e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5656353
Reviewed-by: Vaibhav Raheja <vraheja@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Vaibhav Raheja <vraheja@chromium.org>
Tested-by: Vaibhav Raheja <vraheja@chromium.org>