shill: Network: Receiving DHCP option 108 won't cause Network to stop
Pre this patch, when DHCP option 108 is received and there is no IPv6
configuration available, the Network class only waits for SLAAC if it is
in Configuring state, elsewise will Stop the Network explicitly. With
this patch Network will never Stop on option 108 and will continue to
wait for IPv6 provisioning.

BUG=b:359931692
TEST=unit

Change-Id: Ia4640f87ee2d2ca71e5116118c9a53e855f4dee9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5839065
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Taoyu Li <taoyl@chromium.org>
Tested-by: Taoyu Li <taoyl@chromium.org>