init: Change has-splash-screen default to true
The display can only be active at this point in the boot process if
it has been initialized in the firmware. If so, we do not need to
render the logo again from user mode.

While we have full control over the firmware in all ChromeOS devices,
that's not the case for Flex devices. To address this, we added a
configuration flag using `cros_config` in commit
3862248e5da6a021e8d6 ("init: Fix condition to accommodate Flex
devices").

After uprev'ing to newer kernels and adding support for seamless
display handoff, we are ready to deploy the feature to more
platforms. Now, instead of changing the configuration for every
single device so that has-splash-screen returns true, it would be
more efficient to assume it's true unless we return false (which will
only be the case in Flex). As mentioned previously, this is safe
because of the other condition in the if statement, and because
we know the display could only be active at this point if it had been
initialized by the firmware, which we control.

BUG=b:302505216, b:364384225
TEST=Manually check condition on Flex and ChromeOS devices

Change-Id: I8847b4a819d3801fafb36b93fd7dabe337f80a2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5840247
Commit-Queue: Paz Zcharya <pazz@google.com>
Auto-Submit: Paz Zcharya <pazz@google.com>
Tested-by: Paz Zcharya <pazz@google.com>
Reviewed-by: Subrata Banik <subratabanik@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Commit-Queue: ChromeOS Auto Runner <chromeos-auto-runner@chromeos-bot.iam.gserviceaccount.com>