camera: usb: Add quirk for HDMI input devices
Some HDMI input devices will block in VIDIOC_DQBUF when no HDMI video
source is connected. For those devices we prefer to return an error for
those capture requests instead of hanging. Hanging causes the camera
service to restart due to the monitors in the HAL adapter not getting
kicked.

Note that using the existing kQuirkRestartOnTimeout didn't work since
that will try to start/stop the stream when a timeout occurs and that
won't kick the monitor because it will just timeout again.

BUG=b:354766714
TEST=Manually tested with Lattice HDMI input

Change-Id: Ic0eccc1234c211f92d1732ba72aba18a1f23c883
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5838875
Commit-Queue: Pablo Ceballos <pceballos@chromium.org>
Tested-by: Pablo Ceballos <pceballos@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>